refactor adventure map ui
