+++ src/features/adventure/ui/zoneUI.js
@@ -1,33 +1,38 @@
-import { S } from '../../../shared/state.js';
 import { ZONES } from '../data/zones.js';
+import { getAdventure } from '../selectors.js';
+import { selectZone, selectArea, updateActivityAdventure } from '../logic.js';
 
-export function updateZoneButtons(selectZone) {
+export function updateZoneButtons() {
   const zoneContainer = document.getElementById('zoneButtons');
-  if (!zoneContainer || !S.adventure) return;
+  const adventure = getAdventure();
+  if (!zoneContainer || !adventure) return;
   zoneContainer.innerHTML = '';
   ZONES.forEach((zone, index) => {
-    if (index < (S.adventure.zonesUnlocked || 1)) {
+    if (index < (adventure.zonesUnlocked || 1)) {
       const button = document.createElement('button');
-      button.className = 'btn zone-btn' + (index === (S.adventure.selectedZone || 0) ? ' active' : '');
+      button.className = 'btn zone-btn' + (index === (adventure.selectedZone || 0) ? ' active' : '');
       button.textContent = zone.name;
-      button.onclick = () => selectZone(index);
+      button.onclick = () => {
+        if (selectZone(index)) updateActivityAdventure();
+      };
       zoneContainer.appendChild(button);
     }
   });
 }
 
-export function updateAreaGrid(selectArea) {
+export function updateAreaGrid() {
   const areaContainer = document.getElementById('areaGrid');
-  if (!areaContainer || !S.adventure) return;
-  const currentZone = ZONES[S.adventure.selectedZone || 0];
+  const adventure = getAdventure();
+  if (!areaContainer || !adventure) return;
+  const currentZone = ZONES[adventure.selectedZone || 0];
   if (currentZone && currentZone.areas) {
     areaContainer.innerHTML = '';
     currentZone.areas.forEach((area, index) => {
       const button = document.createElement('button');
-      const areaKey = `${S.adventure.selectedZone}-${index}`;
-      const isUnlocked = S.adventure.unlockedAreas[areaKey] || false;
-      const isActive = index === (S.adventure.selectedArea || 0);
-      const progress = S.adventure.areaProgress[areaKey] || { kills: 0, bossDefeated: false };
+      const areaKey = `${adventure.selectedZone}-${index}`;
+      const isUnlocked = adventure.unlockedAreas[areaKey] || false;
+      const isActive = index === (adventure.selectedArea || 0);
+      const progress = adventure.areaProgress[areaKey] || { kills: 0, bossDefeated: false };
 
       button.className = 'btn area-btn' + (isActive ? ' active' : '') + (!isUnlocked ? ' disabled' : '');
       button.textContent = area.name;
@@ -36,7 +41,9 @@ export function updateAreaGrid(selectArea) {
       if (isUnlocked) {
         const statusText = progress.bossDefeated ? ' ✓' : (progress.kills >= area.killReq ? ' 👹' : ` (${progress.kills}/${area.killReq})`);
         button.textContent += statusText;
-        button.onclick = () => selectArea(index);
+        button.onclick = () => {
+          if (selectArea(index)) updateActivityAdventure();
+        };
       } else {
         button.textContent += ' 🔒';
       }
